A better implementation of JumpPointRefinementPolicy. Only increases depth if branching factor of a node is > 1. NB: still not sufficiently general. it doesn't differentiate between a node whose branching factor is actually > 1 (i.e. could not prove that all neighbours lead to dead ends) and nodes whose branching factor appears > 1 but which have neighbours that lead to dead ends.
Other stuff:
Jump.h: Direction::NONE is now enum value 0.

JumpPointLocator: added computeSuccessors method
which is made up of ::computeNatural and ::computeForced.
(useful if we need to identify subsets of directions to search in for a
particular node).
NB: ::computeForced used to live in OnlineJumpPointLocator but now resides here.

JumpPointsExpansionPolicy: added a value to track recursion depth in the future.